+++ gzap/CHANGELOG.md
@@ -11,6 +11,8 @@
   - If more progression items exist than free locations, it will skip filler
     placement (rather than crashing), since generation may still succeed in
     multiworld where the extra items can be displace filler from other worlds.
+  - Restarting uzdoom to load a different game while leaving the client open no
+    longer sends out checks from the previous game.
 - For logic developers:
   - `destroy-on-spawn` directive in `GZAPRC`.
   - AP can now put multiple copies of a `key` in the pool and keep track of how